added Security features

+++ onlineshop-spring-mvc-webapp/src/main/java/com/sci/onlineshop/dao/ProductDao.java
@@ -13,9 +13,9 @@ public interface ProductDao {
     
     void editProduct(Product product);
 
-    Product getProductById(int id);
+    Product getProductById(long id);
 
     List<Product> getAllProducts();
 
-    void deleteProduct(int id);
+    void deleteProduct(long id);
 }
